Route changing

+++ src/Commands/DeleteTempUploads.php
@@ -7,7 +7,8 @@
     use Util;
 
     /**
-     * Reactables delete temporary uploads command.
+     * Reactables delete temporary uploads Firefly command.\
+     * Usage: `reactables:delete-temp-uploads`
      * @category Command
      * @package glowieframework/reactables
      * @author Glowie
@@ -21,16 +22,18 @@
          * The command script.
          */
         public function run(){
+            // Print title
+            $this->print('<bg="yellow"><color="black">Reactables</color></bg> ', false);
+
             // Get temp directory
             $dir = Config::get('reactables.tmp_path', Util::location('storage/reactables'));
             if(!is_writable($dir)) throw new FileException('Directory "' . $dir . '" is not writable, please check your chmod settings');
 
             // Delete files
-            foreach (Util::getFiles($dir . '/*') as $filename) unlink($filename);
-
-            // Return result
-            $this->print('<bg="yellow"><color="black">Reactables</color></bg> ', false);
-            $this->success("Temporary uploads deleted successfully!");
+            if($this->confirm('<color="yellow">Delete temporary uploads?</color>')){
+                foreach (Util::getFiles($dir . '/*') as $filename) unlink($filename);
+                $this->success("Temporary uploads deleted successfully!");
+            }
         }
 
     }
